rust-vex: implement SSE scalar float ops and fix XMM register handling
- Add VFAddS, VFSubS, VFMulS, VFDivS operations for SSE scalar-in-vector
  float ops (ADDSS, SUBSS, MULSS, DIVSS, etc.)
- Map Iop_Add32F0x4, Iop_Div32F0x4, etc. to the new VFxxx operations
- Fix AMD64 XMM register offsets to match archinfo/pyvex:
  - xmm0 at offset 224 (was 256)
  - xmm1 at offset 256 (was 272)
  - etc.
- Change get_register/set_register to use u128 instead of u64 to
  properly handle 128-bit XMM registers
- Add unit tests for VFAddS and VFDivS operations

+++ native/angr/src/engine.rs
@@ -269,7 +269,8 @@ impl RustVEXEngine {
     }
 
     /// Get a register value.
-    pub fn get_register(&self, name: &str) -> PyResult<u64> {
+    /// Returns u128 to handle XMM and other large registers (Python BigInt handles this).
+    pub fn get_register(&self, name: &str) -> PyResult<u128> {
         let arch = arch_from_name(&self.arch_name).unwrap();
         let offset = arch.register_offset(name).ok_or_else(|| {
             PyValueError::new_err(format!("unknown register: {}", name))
@@ -282,15 +283,16 @@ impl RustVEXEngine {
             return Err(PyValueError::new_err("register out of bounds"));
         }
 
-        let mut value: u64 = 0;
+        let mut value: u128 = 0;
         for i in 0..size {
-            value |= (self.registers[offset + i] as u64) << (i * 8);
+            value |= (self.registers[offset + i] as u128) << (i * 8);
         }
         Ok(value)
     }
 
     /// Set a register value.
-    pub fn set_register(&mut self, name: &str, value: u64) -> PyResult<()> {
+    /// Accepts u128 to handle XMM and other large registers (Python BigInt handles this).
+    pub fn set_register(&mut self, name: &str, value: u128) -> PyResult<()> {
         let arch = arch_from_name(&self.arch_name).unwrap();
         let offset = arch.register_offset(name).ok_or_else(|| {
             PyValueError::new_err(format!("unknown register: {}", name))
